+++ src/Classes/QREncoder/QREncoder.h
@@ -33,6 +33,6 @@
 }
 
 + (UIImage *)encode:(NSString *)str;
-+ (UIImage *)encode:(NSString *)str size:(int)size correctionLevel:(QRCorrectionLevel)level;
++ (UIImage *)encode:(NSString *)str size:(NSInteger)size correctionLevel:(QRCorrectionLevel)level;
 
 @end
